Fetch available dates for talk form from db
* Move "home" stuff into it's own component
* Fix "children must have unique keys" issue
* Fix date formatting
