+++ functions/scheduled-events/read-all.js
@@ -1,7 +1,7 @@
 /* Import faunaDB sdk */
 const process = require('process')
 
-const { query, Client } = require('faunadb')
+const { query: q, Client } = require('faunadb')
 
 const client = new Client({
   secret: process.env.FAUNADB_SERVER_SECRET,
@@ -10,13 +10,13 @@ const client = new Client({
 const handler = async () => {
   console.log('Function `read-all` invoked')
   return client
-    .q(query.Paginate(query.Match(query.Index('all_items'))))
+    .query(q.Paginate(q.Match(q.Index('all_scheduled_events'))))
     .then((response) => {
       const itemRefs = response.data
       // create new query out of item refs. http://bit.ly/2LG3MLg
-      const getAllItemsDataQuery = itemRefs.map((ref) => query.Get(ref))
+      const getAllItemsDataQuery = itemRefs.map((ref) => q.Get(ref))
       // then query the refs
-      return client.q(getAllItemsDataQuery).then((ret) => ({
+      return client.query(getAllItemsDataQuery).then((ret) => ({
         statusCode: 200,
         body: JSON.stringify(ret),
       }))
